UPD fix patients list styling
- Fixes patients list (both general patient list and waitlist)
  styling according to guidelines:
    - Border should start after avatar/image.
    - Proper alignment.
    - Main complain in the waitlist.
    - Nationality displayed with code title, not code.
