+++ frontend/local/src/containers/waitlist/detail/index.js
@@ -10,12 +10,8 @@ import { ReactComponent as ComplaintIcon } from "shared/icons/complaint.svg"
 import { ReactComponent as DiagnosisIcon } from "shared/icons/diagnosis.svg"
 import { ReactComponent as MedicalDataIcon } from "shared/icons/vitalsigns.svg"
 import { ReactComponent as TherapyIcon } from "shared/icons/therapy.svg"
-//import { ReactComponent as LaboratoryIcon } from "shared/icons/laboratory.svg"
-//import { ReactComponent as NegativeIcon } from "shared/icons/negative.svg"
-//import { ReactComponent as PositiveIcon } from "shared/icons/positive.svg"
 
 import { joinPaths } from "shared/utils"
-import { fetchCode } from "shared/modules/codes"
 import {
     RESOURCE_PATIENT_IDENTIFICATION,
     RESOURCE_VITAL_SIGNS,
@@ -36,55 +32,8 @@ import AddDiagnosis from "./add-diagnosis"
 import Remove from "./remove"
 import Spinner from "shared/containers/spinner"
 import VitalSignCard from "shared/containers/vitalSign"
+import { CodeTitle } from "shared/containers/codes"
 import { open, COLOR_SUCCESS } from "shared/modules/alert"
-
-class CodeTitle extends React.Component {
-    constructor(props) {
-        super(props)
-        this.loadCode = this.loadCode.bind(this)
-        if (props.codeId) {
-            this.loadCode(props.categoryId, props.codeId)
-        }
-
-        this.componentWillUnmount = this.componentWillUnmount.bind(this)
-        this.state = { loading: true, title: "" }
-    }
-
-    loadCode(categoryId, codeId) {
-        this.props
-            .fetchCode(categoryId, codeId)
-            .then(code => {
-                if (!this.unmounted) {
-                    this.setState({
-                        loading: false,
-                        failed: false,
-                        title: code.title
-                    })
-                }
-            })
-            .catch(ex => {
-                this.setState({ loading: false, failed: true })
-            })
-    }
-
-    componentWillUnmount() {
-        this.unmounted = true
-    }
-
-    render() {
-        if (this.state.loading) {
-            return <span>...</span>
-        }
-
-        if (this.state.failed) {
-            return <span>Failed to fetch title!</span>
-        }
-
-        return <span>{this.state.title}</span>
-    }
-}
-
-CodeTitle = connect(state => ({}), { fetchCode })(CodeTitle)
 
 class InConsultation extends React.Component {
     constructor(props) {
