+++ frontend/local/src/containers/waitlist/style.scss
@@ -11,7 +11,7 @@
 
     .part {
         margin: 50px 0;
-        padding: 30px 20px;
+        padding: 10px 20px;
     }
 
     .now,
